api/requirements.txt
azure-functions
redis
psycopg2-binary
stripe
anthropic
